Gate werkt niet

+++ app/Http/Controllers/DogController.php
@@ -8,6 +8,7 @@ use App\Models\Breed;
 use Illuminate\Support\Str;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class DogController extends Controller
 {
@@ -40,6 +41,9 @@ class DogController extends Controller
     }
 
     public function edit (Dog $dog) {
+        if (! Gate::allows('ownership', Auth::user())) {
+            abort(403);
+        }
         $breeds=Breed::all();
         return Inertia::render('Dogs/EditDog', [
             'breeds' => $breeds,
@@ -48,6 +52,9 @@ class DogController extends Controller
     }
 
     public function update (Request $request, Dog $dog) {
+        if (! Gate::allows('ownership', $dog)) {
+            abort(403);
+        }
         $request->validate([
             'breed_id' => 'required',
             'date_of_birth' => 'required',
@@ -65,6 +72,9 @@ class DogController extends Controller
     }
 
     public function destroy (Dog $dog) {
+        if (! Gate::allows('ownership', $dog)) {
+            abort(403);
+        }
         $dog->ownerships()->detach(Auth::user()); //TO DO: is dit nodig? Zo ja, zelfde voor memberships
         $dog->delete();
         return redirect()->route('dashboard');
